requirements.txt
boto3==1.4.5
Flask==0.12.2
Flask-Cors==3.0.2
flask-oidc==1.1.1
flask_restful
flask-swagger==0.2.13
Flask-SQLAlchemy==2.2
pymysql==0.7.11
alembic==0.9.3
pyBigWig==0.3.7
certifi>3017
research-auth==0.3.5
